+++ src/tests/ha_addon_sunsynk_multi/test_driver.py
@@ -7,7 +7,6 @@ from ha_addon_sunsynk_multi.driver import STATE, init_driver
 from ha_addon_sunsynk_multi.options import OPT
 from sunsynk.pysunsynk import PySunsynk
 from sunsynk.solarmansunsynk import SolarmanSunsynk
-from sunsynk.usunsynk import USunsynk
 
 
 def test_init() -> None:
@@ -28,17 +27,6 @@ def test_init() -> None:
     ist = STATE[0].connector[0]
     assert isinstance(ist, PySunsynk)
     assert ist.port == inv_port
-    # assert STATE[0].inv == PySunsynk(port=inv_port, state=STATE[0].inv_state)
-
-    AInverter.connectors.clear()
-    OPT.driver = "umodbus"
-    OPT.load_dict(inv_option)
-    init_driver(OPT)
-    assert len(STATE) == 1
-    ist = STATE[0].connector[0]
-    assert isinstance(ist, USunsynk)
-    assert ist.port == inv_port
-    # assert STATE[0].inv == USunsynk(port=inv_port, state=STATE[0].inv_state)
 
     AInverter.connectors.clear()
     inv_option = {
@@ -53,6 +41,3 @@ def test_init() -> None:
     assert isinstance(ist, SolarmanSunsynk)
     assert ist.port == inv_port
     assert ist.dongle_serial_number == 101
-    # assert STATE[0].inv == SolarmanSunsynk(
-    #     port=inv_port, state=STATE[0].inv_state, dongle_serial_number=101
-    # )